requirements.txt
DateTime==5.5
icmplib==3.0.4
narwhals==1.43.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
plotly==6.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
speedtest-cli==2.1.3
tzdata==2025.2
zope.interface==7.2
